Refactoring docservice to allow multiple spaces with only one docservice instance

+++ applications/space/server/src/org/monet/space/kernel/agents/AgentFilesystem.java
@@ -87,7 +87,9 @@ public class AgentFilesystem {
 	}
 
 	public static void removeDirContent(File directory) {
-		for (File file : directory.listFiles()) {
+		File[] files = directory.listFiles();
+		if (files == null) return;
+		for (File file : files) {
 			if (file.isDirectory())
 				removeDir(file.getAbsolutePath());
 			else
